+++ lib/domain/model/chat.dart
@@ -41,6 +41,7 @@ class Chat implements Comparable<Chat> {
     this.members = const [],
     this.kindIndex = 0,
     this.isHidden = false,
+    this.isArchived = false,
     this.muted,
     this.directLink,
     PreciseDateTime? createdAt,
@@ -76,13 +77,19 @@ class Chat implements Comparable<Chat> {
   /// Kind of this [Chat].
   int kindIndex;
 
+  /// Get [ChatKind] of this [Chat].
   ChatKind get kind => ChatKind.values[kindIndex];
+
+  /// Sets the [ChatKind] of this [Chat].
   set kind(ChatKind chatKind) {
     kindIndex = chatKind.index;
   }
 
   /// Indicator whether this [Chat] is hidden by the authenticated [MyUser].
   bool isHidden;
+
+  /// Indicator whether this [Chat] is archived by the authenticated [MyUser].
+  bool isArchived;
 
   /// Mute condition of this [Chat] for the authenticated [MyUser].
   ///
@@ -290,6 +297,7 @@ class Chat implements Comparable<Chat> {
       kindIndex: kindIndex,
       muted: muted,
       isHidden: isHidden,
+      isArchived: isArchived,
       directLink: directLink,
       createdAt: createdAt,
       updatedAt: updatedAt,
@@ -318,6 +326,7 @@ class Chat implements Comparable<Chat> {
         const ListEquality().equals(members, other.members) &&
         kindIndex == other.kindIndex &&
         isHidden == other.isHidden &&
+        isArchived == other.isArchived &&
         muted == other.muted &&
         directLink == other.directLink &&
         createdAt == other.createdAt &&
@@ -335,13 +344,14 @@ class Chat implements Comparable<Chat> {
   }
 
   @override
-  int get hashCode => Object.hash(
+  int get hashCode => Object.hashAll([
     id,
     avatar,
     name,
     members,
     kindIndex,
     isHidden,
+    isArchived,
     muted,
     directLink,
     createdAt,
@@ -356,7 +366,7 @@ class Chat implements Comparable<Chat> {
     ongoingCall,
     favoritePosition,
     membersCount,
-  );
+  ]);
 }
 
 /// Member of a [Chat].