+++ lib/domain/service/chat.dart
@@ -54,6 +54,9 @@ class ChatService extends DisposableService {
   /// Returns the reactive map of the currently paginated [RxChat]s.
   RxObsMap<ChatId, RxChat> get paginated => _chatRepository.paginated;
 
+  /// Returns the reactive map of the archived chats paginated [RxChat]s.
+  RxObsMap<ChatId, RxChat> get archived => _chatRepository.archived;
+
   /// Returns the current reactive map of all [RxChat]s available.
   RxObsMap<ChatId, RxChat> get chats => _chatRepository.chats;
 
@@ -186,6 +189,12 @@ class ChatService extends DisposableService {
     }
 
     return _chatRepository.hideChat(id);
+  }
+
+  /// Archives or unarchives the specified [Chat] for the authenticated [MyUser].
+  Future<void> archiveChat(ChatId id, bool archive) {
+    Log.debug('archiveChat($id, $archive)', '$runtimeType');
+    return _chatRepository.archiveChat(id, archive);
   }
 
   /// Adds an [User] to a [Chat]-group by the authority of the authenticated