+++ lib/ui/page/home/tab/chats/controller.dart
@@ -57,6 +57,7 @@ import '/provider/gql/exceptions.dart'
         HideChatException,
         JoinChatCallException,
         RemoveChatMemberException,
+        ToggleChatArchivationException,
         ToggleChatMuteException,
         UnfavoriteChatException;
 import '/routes.dart';
@@ -85,6 +86,9 @@ class ChatsTabController extends GetxController {
   /// Reactive list of sorted [Chat]s.
   final RxList<ChatEntry> chats = RxList();
 
+  /// Reactive list of sorted archived [Chat]s.
+  final RxList<ChatEntry> archived = RxList();
+
   /// [SearchController] for searching the [Chat]s, [User]s and [ChatContact]s.
   final Rx<SearchController?> search = Rx(null);
 
@@ -93,6 +97,9 @@ class ChatsTabController extends GetxController {
 
   /// Indicator whether [search]ing is active.
   final RxBool searching = RxBool(false);
+
+  /// Indicator whether chat archive viewing is active.
+  final RxBool archivedOnly = RxBool(false);
 
   /// [ScrollController] to pass to a [Scrollbar].
   final ScrollController scrollController = ScrollController();
@@ -152,6 +159,9 @@ class ChatsTabController extends GetxController {
 
   /// Subscription for the [ChatService.paginated] changes.
   late final StreamSubscription _chatsSubscription;
+
+  /// Subscription for the [ChatService.archived] changes.
+  late final StreamSubscription _archivedSubscription;
 
   /// Subscription for [SearchController.chats], [SearchController.users] and
   /// [SearchController.contacts] changes updating the [elements].
@@ -269,6 +279,31 @@ class ChatsTabController extends GetxController {
       }
     });
 
+    _archivedSubscription = _chatService.archived.changes.listen((event) {
+      switch (event.op) {
+        case OperationKind.added:
+          final entry = ChatEntry(event.value!, chats.sort);
+          archived.add(entry);
+          archived.sort();
+          break;
+
+        case OperationKind.removed:
+          archived.removeWhere((e) {
+            if (e.chat.value.id == event.key) {
+              e.dispose();
+              return true;
+            }
+
+            return false;
+          });
+          break;
+
+        case OperationKind.updated:
+          archived.sort();
+          break;
+      }
+    });
+
     if (_chatService.status.value.isSuccess) {
       SchedulerBinding.instance.addPostFrameCallback(
         (_) => _ensureScrollable(),
@@ -299,6 +334,7 @@ class ChatsTabController extends GetxController {
       data.dispose();
     }
     _chatsSubscription.cancel();
+    _archivedSubscription.cancel();
     _statusSubscription?.cancel();
 
     _searchSubscription?.cancel();
@@ -396,6 +432,46 @@ class ChatsTabController extends GetxController {
     } catch (e) {
       MessagePopup.error(e);
       rethrow;
+    }
+  }
+
+  /// Archives or unarchives the specified [Chat] identified by the provided
+  /// [id].
+  Future<void> archiveChat(ChatId id, bool archive) async {
+    try {
+      await _chatService.archiveChat(id, archive);
+    } on ToggleChatArchivationException catch (e) {
+      MessagePopup.error(e);
+    } on UnfavoriteChatException catch (e) {
+      MessagePopup.error(e);
+    } catch (e) {
+      MessagePopup.error(e);
+      rethrow;
+    }
+  }
+
+  /// Archives or unarchives the [selectedChats].
+  Future<void> archiveChats(bool archive) async {
+    selecting.value = false;
+
+    router.navigation.value = !selecting.value;
+    router.navigator.value = null;
+
+    try {
+      await Future.wait(
+        selectedChats.map(
+          (chatId) => _chatService.archiveChat(chatId, archive),
+        ),
+      );
+    } on ToggleChatArchivationException catch (e) {
+      MessagePopup.error(e);
+    } on UnfavoriteChatException catch (e) {
+      MessagePopup.error(e);
+    } catch (e) {
+      MessagePopup.error(e);
+      rethrow;
+    } finally {
+      selectedChats.clear();
     }
   }
 
@@ -687,6 +763,7 @@ class ChatsTabController extends GetxController {
               e.chat.value.ongoingCall == null &&
               e.chat.value.favoritePosition != null &&
               !e.chat.value.isHidden &&
+              !e.chat.value.isArchived &&
               !e.hidden.value,
         )
         .toList();
@@ -771,6 +848,11 @@ class ChatsTabController extends GetxController {
       MessagePopup.error('err_data_transfer'.l10n);
       rethrow;
     }
+  }
+
+  /// Toggles the [archivedOnly].
+  void toggleArchive() {
+    archivedOnly.value = !archivedOnly.value;
   }
 
   /// Enables and initializes or disables and disposes the [search].