+++ test/unit/chat_attachment_test.dart
@@ -95,6 +95,9 @@ void main() async {
     graphQlProvider.recentChatsTopEvents(3),
   ).thenAnswer((_) => const Stream.empty());
   when(
+    graphQlProvider.recentChatsTopEvents(3, archived: true),
+  ).thenAnswer((_) => const Stream.empty());
+  when(
     graphQlProvider.incomingCallsTopEvents(3),
   ).thenAnswer((_) => const Stream.empty());
   when(graphQlProvider.keepOnline()).thenAnswer((_) => const Stream.empty());
@@ -135,6 +138,7 @@ void main() async {
         last: null,
         before: null,
         noFavorite: anyNamed('noFavorite'),
+        archived: anyNamed('archived'),
         withOngoingCalls: anyNamed('withOngoingCalls'),
       ),
     ).thenAnswer((_) => Future.value(RecentChats$Query.fromJson(recentChats)));
@@ -249,6 +253,7 @@ void main() async {
         last: null,
         before: null,
         noFavorite: anyNamed('noFavorite'),
+        archived: anyNamed('archived'),
         withOngoingCalls: anyNamed('withOngoingCalls'),
       ),
     ).thenAnswer((_) => Future.value(RecentChats$Query.fromJson(recentChats)));
@@ -354,6 +359,7 @@ final chatData = {
   'members': {'nodes': [], 'totalCount': 0},
   'kind': 'GROUP',
   'isHidden': false,
+  'isArchived': false,
   'muted': null,
   'directLink': null,
   'createdAt': '2021-12-15T15:11:18.316846+00:00',